chore: move files to comply with package layout conventions
